Extrcated ReadPipe and WritePipe from process.[ch]
